V2 (#98)
Closes #5 

*Description of changes:*

Provide a major version 2 version of the Java Caching Library.

Version 2 uses the Java SDK v2 under the hood. The library should be a
drop-in replacement, unless you currently provide a v1 client to the
caching library at creation. In that case, you will have to start
passing a v2 client instead.

Version 1 and version 2 will be maintained side by side in separate
branches.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.